requirements.txt
astrapy==2.2.1
beautifulsoup4==4.14.3
feedparser==6.0.12
googlesearch-python==1.3.0
huggingface-hub==1.7.2
langchain==1.2.12
langchain-core==1.2.22
langchain-groq==1.1.2
langchain-openai==1.1.12
numpy==2.4.3
openai==2.29.0
playwright==1.58.0
psycopg2-binary==2.9.9
pypdf==5.9.0
python-dotenv==1.0.0
requests==2.33.1
torch==2.10.0
transformers==5.3.0
